do not fail silently in case of eslint error
when upgrading from eslint 2 to 8,
it could not find the airbnb config anymore.

But the git pre-commit hook did not say anything
(and did not report any linting errors)

An error code of `1` means that there are lint errors.
An error code of `2` means that eslint could not run.
(see https://eslint.org/docs/user-guide/command-line-interface#exit-codes)

throwing this exception makes the hook abort and displays the error
message.

in our case:

```
ESLint couldn't find the config "airbnb/legacy" to extend from. Please check that the name of the config is correct.
The config "airbnb/legacy" was referenced from the config file in "/var/www/allegro/Ressource/CodeStandards/legacy.eslintrc.json".
```