avoid eslint error
when upgrading from eslint 2 to 8 eslint would not run anymore:

```
ESLint couldn't find the config "airbnb/legacy" to extend from. Please check that the name of the config is correct.
The config "airbnb/legacy" was referenced from the config file in "/var/www/allegro/Ressource/CodeStandards/legacy.eslintrc.json".
```

we run it with the `--config` option to specify the eslint config file

but that file is in another directory
and it seems that eslint now looks for the extended config relative to
that file.

=> copy the config file in this tool's directory structure,
so that eslint can find the extended config that is defined in
its `package.json` file.