also copy extended eslint config files
in a previous commit we needed to copy config files locally
otherwise if a config extended for example "airbnb"
and airbnb was installed in the code checking repo but not in the
directory hierarchy of the config file eslint would not find it.

but if a config file extends another local file we need to copy that
file too (potentially recursively).

in our case a config file extended this:

```
    "extends": "./legacy.eslintrc.json",
```